+++ cansocket/examples/async_tokio.rs
@@ -6,6 +6,7 @@
 async fn main() -> Result<(), cansocket::Error> {
     let socket = cansocket::Socket::new("vcan0")?;
     socket.set_recv_own_msgs(true)?;
+    socket.set_protocol_errors(true)?;
 
     println!("Sending a frame.");
     let id = embedded_can::StandardId::new(0x123).unwrap();
@@ -13,7 +14,7 @@ async fn main() -> Result<(), cansocket::Error> {
     socket.send(&frame).await?;
 
     println!("Receiving a frame.");
-    let frame = socket.recv().await?;
+    let frame = socket.recv().await;
     println!("Received: {:?}", frame);
 
     Ok(())