+++ cansocket/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Use libc constants for error frame checking
 - Fix `Socket` doc test
 - Added option to enable protocol errors
+- Added async example with smol
 
 ## v0.1.0